add validation to delete button

+++ Project/views/crowd_project.py
@@ -1,4 +1,4 @@
-from django.db.models import Avg
+from django.db.models import Avg, Sum
 from django.forms import modelformset_factory
 from django.shortcuts import render, redirect, get_object_or_404
 from django.http import HttpResponse
@@ -7,8 +7,9 @@ from django.template import RequestContext
 from Project.forms.project_form import ProjectForm, ImageForm
 from Project.forms.project_form import TagForm
 from Project.models.category import Category
-from Project.models.user_project import UserProject, ProjectRate
+from Project.models.user_project import UserProject, ProjectRate, ProjectDonation
 from Project.models.featured_project import FeaturedProject
+from User.models import User
 
 from Project.models.project_picture import ProjectPicture
 from Project.models.tag import Tag
@@ -169,9 +170,18 @@ def project_form(request):
 def delete(request, project_id):
     if not request.user.is_authenticated:
         return redirect('login')
-
+    
     project = get_object_or_404(UserProject, id=project_id)
-    project.delete()
+    currentAmout=ProjectDonation.objects.filter(project_id=project_id).aggregate(Sum('amount'))["amount__sum"]
+    totalTarget=UserProject.objects.get(id=project_id).total_target
+    
+    if currentAmout  is  None or (currentAmout/totalTarget)*100<25.0:
+        projectToBeDeleted=ProjectDonation.objects.filter(project_id=project_id)
+        for projectDonation in projectToBeDeleted:
+            userBalance = User.objects.values_list('balance', flat=True).filter(id=projectDonation.user_id)
+            userNew=User.objects.filter(id=projectDonation.user_id).update(balance=int(userBalance[0]) + int(projectDonation.amount))
+            
+        project.delete()
     return redirect("list")
 
 
@@ -190,6 +200,5 @@ def featuredProjects(request):
 def category_list(request, category_id):
     if not request.user.is_authenticated:
         return redirect('login')
-
     projects = UserProject.objects.filter(category_id=category_id)
     return render(request, "project/category.html", {'projects': projects})
